requirements_dev.txt
future==0.16.0
pip==9.0.1
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==3.3.0
tox==2.7.0
coverage==4.3.4
Sphinx==1.5.5
PyYAML==3.12
pytest==3.0.7
mockfs==1.0.2
tabulate==0.7.7
click==6.7
configstruct==0.3.1
jmespath==0.9.2
python-dateutil==2.6.0
